requirements.txt
torch
torchvision
Pillow
gradio
matplotlib
scikit-learn
numpy
